==> abbr_decoding/2_predict_abbr.ipynb <==
import pandas as pd
import numpy as np

context_df = pd.read_csv('term_context.csv', sep='\t').dropna()

context_without_term = []

for term, context in context_df[['term', 'context']].iloc:
    context_without_term.append(context.replace(term, ' [TERM] '))
    
context_df['context_without_term'] = context_without_term

from normalizer import normalize_sentences

abbrs = set(context_df['abbreviation'].str.lower())

sentences = normalize_sentences(
    context_df['context_without_term'], 
    except_words=abbrs,
    as_sentences=False
)
clean_sentences = [" ".join([" ".join(word) for word in text]) for text in sentences]
context_df['normalized_context'] = clean_sentences

context_df[context_df.abbreviation == 'Hb'].term.value_counts()

context_df.groupby('abbreviation').term.unique().str.len().sort_values().tail(60)

from sklearn.feature_extraction.text import TfidfVectorizer, CountVectorizer

vectorizer = TfidfVectorizer()
char_vectorizer = CountVectorizer(analyzer='char')

points = vectorizer.fit_transform(context_df['normalized_context']).toarray()
char_points = char_vectorizer.fit_transform(context_df['abbreviation']).toarray()

import joblib

joblib.dump(vectorizer, 'context_vectorizer.joblib')
joblib.dump(char_vectorizer, 'char_vectorizer.joblib')

from sklearn.metrics import accuracy_score, f1_score
from sklearn.ensemble import RandomForestClassifier
from sklearn.linear_model import LogisticRegression
from sklearn.model_selection import train_test_split

df_train, df_test, X_train, X_test, y_train, y_test = train_test_split(
    context_df,
    np.hstack((points, char_points)),
    context_df.term, 
    test_size=0.3
)

clf = RandomForestClassifier()
clf.fit(X_train, y_train) 

joblib.dump(clf, 'decoder_model.joblib')

print('train accuracy:', clf.score(X_train, y_train))
print('train f1_score:', f1_score(y_train, clf.predict(X_train), average='weighted'))

print('test accuracy:', clf.score(X_test, y_test))
print('test f1_score:', f1_score(y_test, clf.predict(X_test), average='weighted'))

df_test = df_test.copy()
df_test['predicted_term'] = clf.predict(X_test)
df_test['proba']= np.max(clf.predict_proba(X_test), axis=1)

df_test[df_test.predicted_term != df_test.term]

- Улучшить задачку какую нибудь

class AbbreviationDecoder:
    def __init__(self):
        self.context_vectorizer = joblib.load('context_vectorizer.joblib')
        self.char_vectorizer = joblib.load('char_vectorizer.joblib')
        self.model = joblib.load('decoder_model.joblib')
    
    def _normalize_sentences(self, sentences):
        sentences = normalize_sentences(
            sentences, 
            except_words=abbrs,
        )
        clean_sentences = [" ".join([" ".join(word) for word in text]) for text in sentences]
        
        return clean_sentences
    
    def _get_features(self, context, abbreviation):
        if isinstance(context, str):
            context = [context]
            abbreviation = [abbreviation]
            
        context = self._normalize_sentences(context)
            
        context_features = vectorizer.transform(context).toarray()
        abbr_features = char_vectorizer.transform(abbreviation).toarray()

        features = np.hstack((context_features, abbr_features))

        return features

    def predict(self, abbreviation, context):
        features = self._get_features(context, abbreviation)
        return self.model.predict(features)

### Evaulation

records_df = pd.read_csv('../data/I10_anamnesis.csv', sep='\t',)

records_df[records_df.RecordEMC.str.findall(' КТ ').str.len() != 0].RecordEMC.iloc[0]

decoder = AbbreviationDecoder()

decoder.predict('ЛП', 'воспалительный изменение')

может ли являтся аббревиатура настоящей аббревиатурой для слова
