==> abbr_decoding/3_text_decoder.ipynb <==
import re

import pandas as pd
import numpy as np

from tqdm.notebook import tqdm
tqdm.pandas()

df = pd.read_csv('../data/abbr_term.csv', sep='\t')

from abbr_decoder import AbbreviationDecoder


def load_extractor():
    import sys
  
    # setting path
    sys.path.append('../abbr_extraction/')
    
    from abbreviation_extractor import load_extractor
    
    sys.path.append('../abbr_decoding/')
    
    return load_extractor()

decoder = AbbreviationDecoder()
extractor = load_extractor()

abbr2term = {abbr.lower(): term for abbr, term in df[['abbreviation', 'term']].iloc if not pd.isna(abbr)}

abbr2term['ад'] = 'артериальное давление'

resolved_pairs = {}

def resolve_abbreviations(text, window=5):
    new_text = []
    words = text.split() + [''] * window
    current_context = [''] * window + words[:window]
    
    for i in range(window, len(words)):
        word = current_context[window]
        
        if extractor.is_abbreviation(word) and extractor._abbr_predictor.predict(word):
            resolved = decoder.predict(word, ' '.join(current_context))[0]
            couldnt_find = (resolved == word)
            
            if couldnt_find:
                clean_word = re.sub('[^а-яА-Яa-zA-Z]', '', word)
                resolved = abbr2term.get(clean_word.lower(), word)
                
            resolved_pairs[word] = (resolved, couldnt_find)
            
            current_context[window] = resolved
            
        new_text.append(current_context.pop(0))
        current_context.append(words[i])
    
    new_text.extend(current_context)
    return " ".join(new_text).strip()

text = 'Постепенное расширение пищевого рациона с увеличением объема жидкой пищи и овощей. Избегать жирной, острой пищи Есть до 6 раз в день мелкими порциями Наблюдение кардиолога и гастроэнтеролога ПРОДОЛЖИТЬ ПОСТОЯННЫЙ ПРИЕМ:  ДИОВАН 80 МГ УТРОМ И ВЕЧЕРОМ ДИЛАТРЕНД 6,25 МГ УТРОМ И ВЕЧЕРОМ АРИФОН-РЕТАРД 1,5 МГ УТРОМ   ПАРИЕТ 20 МГ ЗА 30 МИНУТ ДО ЗАВТРАКА 1 НЕДЕЛЮ, ЗАТЕМ ПО 10 МГ ЗА 30 МИНУТ ДО ЗАВТРАКА 2 МЕСЯЦА  ДЕ-НОЛ ЧЕРЕЗ 1 ЧАС ПОСЛЕ ЗАВТРАКА, ОБЕДА, УЖИНА И НА НОЧЬ 1 МЕСЯЦ  КРЕОН 10000 ЕД В КАЖДЫЙ ПРИЕМ ПИЩИ, А ЕСЛИ УПОТРЕБЛЯЕТСЯ БЕЛКОВАЯ ПИЩА (МЯСО, РЫБА, ТВОРОГ, ЯЙЦО), ТО 25000 ЕД (СУММАРНО ЗА СУТКИ 80000-100000 ЕД)  ПРИ ВЗДУТИИ ЖИВОТА ПЕКСАН-R  '

resolve_abbreviations(text)

## Evaulation

anamnesises_df = pd.read_csv(
    '../../ISC_Analys/all_anamnesises.txt', 
    encoding='cp1251', 
    sep='\t', 
    usecols=['Diagnosis', 'RecordEMC']
).dropna()

anamnesises_df['Letter'] = anamnesises_df.Diagnosis.str[0]

def clear_tag(text):
    return re.sub('<[^>]*>', ' ', text)

anamnesises_df.RecordEMC = anamnesises_df.RecordEMC.map(clear_tag)

anamnesises_df['RecordResolvedAbbreviations'] = anamnesises_df.RecordEMC.progress_apply(resolve_abbreviations)

from normalizer import normalize_sentences

from sklearn.metrics import f1_score
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.feature_extraction.text import TfidfVectorizer

abbrs = set(pd.read_csv('../data/abbr_term.csv', sep='\t')['abbreviation'])

normalized_abbrs_text = normalize_sentences(anamnesises_df.RecordResolvedAbbreviations, except_words=abbrs)

normalized_text = normalize_sentences(anamnesises_df.RecordEMC, except_words=abbrs)

norm_abbr = [" ".join([" ".join(word) for word in text]) for text in normalized_abbrs_text]
norm_text = [" ".join([" ".join(word) for word in text]) for text in normalized_text]

anamnesises_df.RecordEMC.iloc[2]

anamnesises_df.RecordResolvedAbbreviations.iloc[2]

abbr_vectorizer = TfidfVectorizer(max_features=400)
features_abbr = abbr_vectorizer.fit_transform(norm_abbr)

text_vectorizer = TfidfVectorizer(max_features=400)
features_text = text_vectorizer.fit_transform(norm_text)

X_train_abbr, X_test_abbr, X_train_text, X_test_text, y_train, y_test = train_test_split(
    features_abbr,
    features_text,
    anamnesises_df['Letter'],
    test_size=0.3,
    random_state=42
)

def evalulate_model(X_train, X_test):
    clf = LogisticRegression()
    clf.fit(X_train, y_train)
    
    y_pred = clf.predict(X_test)
    print('f1_score:', f1_score(y_pred, y_test, average='weighted'))
    print('accuracy:', clf.score(X_test, y_test))

evalulate_model(X_train_abbr, X_test_abbr)

evalulate_model(X_train_text, X_test_text)

- Классификатор на текстах клиинческих рекомендаций - код мкб, заболевание

len(resolved_pairs)
